tests(nvme): verify bar-mapping and hostmem with a NVMe driver
To be able to verify the functionality provided by hostmem.h, pci.h,
and vfioctl.h, then one approach is to write some drive exercising them.
NVMe devices are great for that as they can be emulated with qemu with
debug-tracing things like mmio operation and commands.
And easy to get a hold of physical devices as well.

Also, this could then also be integrated for high-performance in xNVMe.
However, the intent here is to be able to test the foundational plumbing
of: mmio and dma with PCI attached devices.

Signed-off-by: Simon A. F. Lund <os@safl.dk>