feat(hugepages): add script to inspect and configure hugepages
This script is intended to support the hostmem.h library by setting
up hugepages as needed. It serves a similar role to the xnvme-driver
script, which prepares hugepages for use with DPDK and SPDK, but this
version is focused on hostmem-based memory allocation.

Also provides a convenient way to inspect the current hugepage state.

This is part of the preparation work to improve the xnvme-driver script.

Signed-off-by: Simon A. F. Lund <os@safl.dk>