feat(hostmem_dma): add malloc-like allocator for DMA-capable buffers
Introduces a high-level interface for allocating DMA-capable memory
using a malloc-style API. This marks the culmination of the effort to
split hostmem into focused components.

The goal is to offer a simple and convenient interface for typical
use cases, while still allowing users to access or replace the lower-
level components (heap, hugepages) as needed for customization or
integration with other systems.

Signed-off-by: Simon A. F. Lund <os@safl.dk>