fix(dma): require explicit init/term and add docstrings
The initial implementation used lazy auto-initialization of the heap,
but this made it unnatural for the user to call `hostmem_dma_term()`.
While this could be mitigated by registering a destructor via
`atexit()`, that approach tends to obscure program behavior and invites
bugs.

This change removes lazy init in favor of explicit `hostmem_dma_init()`
and `hostmem_dma_term()` calls, clarifying ownership and lifecycle.

Additionally, this commit adds per-function docstrings and improves
top-level documentation to reflect the current interface.

Signed-off-by: Simon A. F. Lund <os@safl.dk>