refactor(hostmem/dma,nvme): make hostmap_dma re-entrant and parametric
This is done to make the uPCIe library usable for projects with multiple
translation units, as the global state is ofterwise duplicated leading
to repeated and thereby colliding definitions.

nvme_controller_open()
  Now takes the heap as argument and assigns it by argument to extended
  controller struct

nvme_qpair_init()
  Now takes the heap as argument and assigns it by argument to extended
  qpair struct

The following functions are removed: hostmem_dma_init/term() as they
are no longer encapsulating access to a global variable. Thus, simply
drop them.

Additionally, then a couple of comments are removed in the library-descr.

Signed-off-by: Simon A. F. Lund <os@safl.dk>