fix(nvme/controller): set CQID when creating IO submission queue
The CQID was not being set during IO submission queue (SQ) creation.
This commit corrects that oversight.

Additionally, the creation order of the queues has been changed.
Previously, the SQ was created before the completion queue (CQ), but
since the SQ references the CQID, it is more appropriate to create the
CQ first.

Signed-off-by: Simon A. F. Lund os@safl.dk