feat(nvme/request): preallocate page per request for PRP list support
Currently, the NVMe driver primitives support only PRP1, limiting
payloads to a single page. To handle larger payloads, PRP2 and PRP lists
are required.

This change prepares for that by associating and preallocating a
DMA-capable page with each NVMe request, enabling payloads up to 2MB on
systems with 4KB pages.

The request pool is extended with a pointer to DMA-capable memory. Each
request is initialized with a pointer to both the physical and virtual
address of its assigned page.

Allocation and cleanup are handled by:

* nvme_request_pool_init_prps(...)
* nvme_request_pool_term_prps(...)

The intent of the implementation is to support optional or customizable
PRP handling, allowing alternative init/term functions to be used if
needed.

Signed-off-by: Simon A. F. Lund <os@safl.dk>